Confirm that builddocsearchdb is not compilable
https://uk.mathworks.com/matlabcentral/answers/19634-creating-documentation-for-a-compiled-matlab-application

The directory containing this function is on the path, but isn't found by `which`.
I don't know why that is, but it seems you can't use this in a compiled function.
Don't think that documentation is important for a compiled application anyway,
so I'll revert this commit, and remove that section of code.